file restructuring:  created a ui folder for all child components
