+++ src/components/ui/full-stop.tsx
